requirements.txt
pylint==3.3.1
pytest==8.3.3
pytest-cov==6.0.0
pytest-mock==3.14.0
patool==1.15.0
unrar==0.4